Extract mention menu positioning hook
