+++ pr-reports/3137.md
@@ -12,6 +12,7 @@ Fixes mobile `@` user mention suggestions in wave chat composers by keeping the 
 - Repositioned the menu on window, nested-scroll, resize observer, and visual viewport changes so iOS keyboard movement keeps suggestions visible.
 - Guarded missing anchor refs and kept resize observation anchored to the cursor element.
 - Preserved editor focus on option press so tapping a suggested profile can select the mention cleanly.
+- Shared the anchor-based menu placement hook with wave mentions to avoid duplicating the viewport/listener logic.
 - Added focused regression coverage for mobile visual viewport placement, missing anchor safety, and option selection.
 
 ## Frontend Impact
